requirements.txt
python-vlc
pathlib